// angular/src/app/models/DistributorMappingDTO.ts
export interface DistributorMappingDTO {
  customerProductShippedToAddress: string;
  amountCreditedPerItem: string;
  billbackPartnerCustomerNumber: string;
  customerProductShippedToCity: string;
  invoiceShippedDate: string;
  distributorBilledShippedInvoiceNumber: string;
  customerDistributorName: string;
  noMapping: string;
  sgaPXCustomerUnitPrice: string;
  quantityShipped: string;
  rebateClaim: string;
  materialNumber: string;
  billbackSgaPXDistributorCustomerCombination: string;
  customerProductShippedToState: string;
  distributorSgaInvoiceUOM: string;
  distributorInvoiceSgaProductUnitPrice: string;
  customerProductShippedToZipCode: string;
}